backend/requirements.txt
watchdog
python-dotenv
google-generativeai
sentence-transformers
numpy

fastapi
uvicorn
websockets
